기능: Util 기능 추가(#46)
