+++ src/main/java/com/park/muscle/global/infra/redis/RedisConfig.java
@@ -1,0 +1,2 @@
+package com.park.muscle.global.infra.redis;public class RedisConfig {
+}
